backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
psycopg2-binary
sqlparse

